# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Langfuse Admin API
  description: |
    API de administración para Langfuse que permite gestionar organizaciones, proyectos, usuarios y API keys.
    Esta API está diseñada para uso administrativo con autenticación mediante API key.
  version: 1.0.0
  contact:
    name: Langfuse Team
    url: https://langfuse.com
servers:
  - url: http://localhost:3100/api
    description: Servidor local de desarrollo
  - url: https://admin-api.langfuse.com/api
    description: Servidor de producción
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Código de error
        message:
          type: string
          description: Mensaje de error
        details:
          type: object
          description: Detalles adicionales del error (opcional)
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          description: ID único de la API key
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
        public_key:
          type: string
          description: Clave pública (para autenticación)
        display_secret_key:
          type: string
          description: Versión truncada de la clave secreta
        last_used_at:
          type: string
          format: date-time
          description: Fecha de último uso
        note:
          type: string
          description: Nota descriptiva (opcional)
        expires_at:
          type: string
          format: date-time
          description: Fecha de expiración (opcional)
    NewApiKey:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          properties:
            secretKey:
              type: string
              description: Clave secreta completa (solo se muestra una vez)
...
paths:
...
  /projects/{id}/api-keys:
    get:
      summary: Listar API keys
      description: Obtiene las API keys de un proyecto
      operationId: getProjectApiKeys
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
      responses:
        '200':
          description: Lista de API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '404':
          description: Proyecto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Crear API key
      description: Crea una nueva API key para el proyecto
      operationId: createApiKey
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: Descripción opcional
                expiresAt:
                  type: string
                  format: date-time
                  description: Fecha de expiración opcional
      responses:
        '201':
          description: API key creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewApiKey'
        '404':
          description: Proyecto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api-keys/expired:
    get:
      summary: Listar API keys expiradas
      description: Obtiene todas las API keys que han expirado
      operationId: getExpiredApiKeys
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Lista de API keys expiradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Eliminar API keys expiradas
      description: Elimina todas las API keys expiradas
      operationId: cleanupExpiredApiKeys
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: API keys expiradas eliminadas correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación
                  count:
                    type: integer
                    description: Número de API keys eliminadas
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api-keys/{id}:
    get:
      summary: Obtener API key
      description: Obtiene detalles de una API key específica
      operationId: getApiKeyById
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la API key
      responses:
        '200':
          description: Detalles de la API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          description: API key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Eliminar API key
      description: Elimina una API key
      operationId: deleteApiKey
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la API key
      responses:
        '200':
          description: API key eliminada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación
        '404':
          description: API key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api-keys/{id}/regenerate:
    post:
      summary: Regenerar API key
      description: Regenera una API key existente
      operationId: regenerateApiKey
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la API key
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  type: string
                  format: date-time
                  description: Nueva fecha de expiración (opcional)
      responses:
        '200':
          description: API key regenerada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewApiKey'
        '404':
          description: API key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api-keys/{id}/expiration:
    put:
      summary: Actualizar expiración de API key
      description: Actualiza la fecha de expiración de una API key
      operationId: updateApiKeyExpiration
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expiresAt
              properties:
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
                  description: Nueva fecha de expiración (null para eliminar)
      responses:
        '200':
          description: Expiración actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: API key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api-keys/{id}/note:
    put:
      summary: Actualizar nota de API key
      description: Actualiza la nota descriptiva de una API key
      operationId: updateApiKeyNote
      tags: [API Keys]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
                  description: Nueva nota descriptiva
      responses:
        '200':
          description: Nota actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          description: API key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...